Fix compile warning

+++ src/consensus/tx_verify.cpp
@@ -18,7 +18,7 @@
 bool IsFinalTx(const CTransaction &tx, int nBlockHeight, int64_t nBlockTime)
 {
     // MWEB: Check kernel lock heights
-    if (tx.m_mwtx.GetLockHeight() > nBlockHeight)
+    if (tx.m_mwtx.GetLockHeight() > (uint64_t)nBlockHeight)
         return false;
 
     if (tx.nLockTime == 0)
